self in the visit actions is now the expected one
The visit_actions variable is a hash were actions
(who are Proc objects) are stored. So self could be
whatever you want, depending from where you set the
actions. Moreover, in the typical use of this visitor,
self is bind to the visitor class itself, not on
instances. It is a serious issue when you include it
but not when you extend it. This patch solve it.